Update copy and fix layout
